+++ example/src/main/java/im/getsocial/demo/fragment/FriendsFragment.java
@@ -52,6 +52,9 @@ public class FriendsFragment extends BaseFragment {
 
 	private ViewContainer _viewContainer;
 
+	@Nullable
+	private String _newFriend;
+
 	@Override
 	public String getFragmentTag() {
 		return "friends";
@@ -145,6 +148,10 @@ public class FriendsFragment extends BaseFragment {
 		hideLoading();
 	}
 
+	public void setNewFriend(String userId) {
+		_newFriend = userId;
+	}
+
 	class ViewContainer {
 
 		@BindView(R.id.friend_id)
@@ -180,7 +187,13 @@ public class FriendsFragment extends BaseFragment {
 			} else {
 				holder = (ViewHolder) convertView.getTag();
 			}
-			holder.setUser(getItem(position));
+			PublicUser user = getItem(position);
+			holder.setUser(user);
+			if (user.getId().equals(_newFriend)) {
+				convertView.setBackgroundColor(getResources().getColor(android.R.color.holo_green_dark));
+			} else {
+				convertView.setBackgroundColor(getResources().getColor(android.R.color.transparent));
+			}
 			return convertView;
 		}